+++ eurocropsml/acquisition/builder.py
@@ -20,10 +20,6 @@ def build_dataset(
         config: Configuration for acquiring EuroCrops reflectance data.
 
     """
-    final_output_dir = config.raw_data_dir
-    output_dir = config.output_dir
-    local_dir = config.local_dir
-
     vector_data_dir = Settings().data_dir.joinpath(
         "meta_data",
         "vector_data",
@@ -31,6 +27,10 @@ def build_dataset(
 
     config.country_config.post_init(vector_data_dir)
     ct_config = config.country_config
+
+    final_output_dir = config.raw_data_dir.joinpath(ct_config.satellite)
+    output_dir = config.output_dir
+    local_dir = config.local_dir
 
     country = ct_config.country
 
@@ -43,26 +43,30 @@ def build_dataset(
 
     nuts_dir = Settings().data_dir.joinpath("meta_data", "NUTS")
 
-    country_output_dir: Path = output_dir.joinpath(country.replace(" ", "_"))
-    country_output_dir.mkdir(exist_ok=True, parents=True)
+    satellite_output_dir: Path = output_dir.joinpath(country, ct_config.satellite)
+    satellite_output_dir.mkdir(exist_ok=True, parents=True)
 
     logger.info(f"Processing year {ct_config.year} for {country}.")
 
-    collector.acquire_s2_tiles(
+    collector.acquire_sentinel_tiles(
         ct_config,
-        country_output_dir.joinpath("collector"),
+        satellite_output_dir.joinpath("collector"),
         cast(Path, ct_config.shapefile),
         shape_dir_clean,
         config.eodata_dir,
         config.workers,
     )
-    logger.info("Finished step 1: Acquiring list of necessary SAFE-files.")
-    copier.merge_s2_safe_files(
-        cast(list[str], ct_config.bands), country_output_dir, config.workers, local_dir
+    logger.info("Finished step 1: Acquiring list of necessary .SAFE files.")
+    copier.merge_safe_files(
+        ct_config.satellite,
+        cast(list[str], ct_config.bands),
+        satellite_output_dir,
+        config.workers,
+        local_dir,
     )
     if local_dir is not None:
         logger.info(
-            "Finished step 2: Copying SAFE-files to local disk and "
+            "Finished step 2: Copying .SAFE files to local disk and "
             "acquiring list of individual band image paths."
         )
     else:
@@ -70,7 +74,7 @@ def build_dataset(
 
     clipper.clipping(
         ct_config,
-        country_output_dir,
+        satellite_output_dir,
         shape_dir_clean,
         config.workers,
         config.chunk_size,
@@ -81,7 +85,7 @@ def build_dataset(
     logger.info("Finished step 3: Clipping parcels from raster tiles.")
     region.add_nuts_regions(
         ct_config,
-        country_output_dir,
+        satellite_output_dir,
         shape_dir_clean,
         nuts_dir,
         final_output_dir,