+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changes from previous releases are listed below.
 
 ## Upcoming Release
 - Include hidden coverage report files in CI as Github action artifacts _(see #40)_
+- Exit script if automatic acquisition of NUTS files wasn't successful _(see #37)_
+- Add acquisition of Sentinel-1 data _(see #35)_
 - Add documentation of acquiring and processing Sentinel-1 data _(see #38)_
 - Add preprocessing of Sentinel-1 _(see #31)_
 - Renaming and moving of Sentinel-2 bands _(see #42)_